Break components into separate files
