+++ babel.config.js
@@ -11,6 +11,8 @@ module.exports = function(api) {
             'storage/*': './src/storage/*',
             'language': './src/language',
             'language/*': './src/language/*',
+            'settings': './src/components/settings',
+            'settings/*': './src/components/settings/*',
           }
         }
       ]
