+++ src/components/settings/package.json
@@ -1,0 +1,3 @@
+{
+    "name": "settings"
+}
