+++ tsconfig.json
@@ -14,6 +14,8 @@
       "storage/*": ["src/storage/*"],
       "language": ["src/language"],
       "language/*": ["src/language/*"],
+      "settings": ["src/components/settings"],
+      "settings/*": ["src/components/settings/*"],
     }
   },
   "include": ["App.tsx", "src/**/*"],
